requirements.txt
mysql-connector-python==8.3.0
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
langchain==0.1.13
chardet
pinecone
openai
fastapi
uvicorn[standard]
